jwt token 2

+++ controllers/note.js
@@ -50,6 +50,7 @@ exports.createNote = (req, res, next) => {
     title,
     content,
     cover_image: cover_image ? cover_image.path : "",
+    author: req.userId,
   })
     .then((_) => {
       res.status(201).json({
@@ -67,6 +68,7 @@ exports.createNote = (req, res, next) => {
 exports.getNote = (req, res, next) => {
   const { id } = req.params;
   Note.findById(id)
+    .populate("author", "username")
     .then((note) => {
       return res.status(200).json(note);
     })
@@ -82,7 +84,9 @@ exports.deleteNote = (req, res, next) => {
   const { id } = req.params;
   Note.findById(id)
     .then((note) => {
-      unlink(note.cover_image);
+      if (note.cover_image) {
+        unlink(note.cover_image);
+      }
       return Note.findByIdAndDelete(id).then((_) => {
         return res.status(204).json({
           message: "Note deleted",
@@ -120,7 +124,9 @@ exports.updateNote = (req, res, next) => {
       note.title = title;
       note.content = content;
       if (cover_image) {
-        unlink(note.cover_image);
+        if (note.cover_image) {
+          unlink(note.cover_image);
+        }
         note.cover_image = cover_image.path;
       }
       return note.save();
